feat: WebUI改修
- useSystemManagerClient.ts:
  - APIログ、コールAPI、システムマネージャーハンドルの型を別ファイルからインポートするように変更。

- SettingPage.tsx:
  - トグルコンポーネントを別ファイルからインポートし、内部で定義していたトグル関数を削除。
  - 日本語の説明文を英語に翻訳。

- SlamPage.tsx:
  - コンテナステータスカードのタイトルを「SLAM Container」から「SLAM」に変更。
  - サービスコントロールカードのタイトルを「SLAM Control」から「Control」に変更。
  - 日本語のラベルを英語に翻訳。

- SystemPage.tsx:
  - サービス管理のセクションを改善し、サービスの開始と停止のロジックを簡素化。
  - 診断詳細の表示を改善し、DiagnosticsTableコンポーネントを使用。

- TopPage.tsx:
  - アラートセクションで診断ステータスを表示するためにDiagnosticsTableを使用。

- WaypointNavPage.tsx:
  - コンテナステータスのタイトルを「Navigation Container」から「Navigation」に変更。
  - 各セクションのタイトルを英語に翻訳。

- ROS関連の型定義を整理し、ファイルを分割。
  - namespaces.ts、schemas.ts、services.ts、topics.tsを新規作成し、関連する型や定数を整理。
