+++ mg_ui/mg_web_ui/frontend/src/components/sections/ApiLogSection.tsx
@@ -1,10 +1,1 @@
-import { ApiLog } from "../../hooks/useSystemManagerClient";
-import ApiLogPanel from "../ApiLogPanel";
-
-interface ApiLogSectionProps {
-  logs: ApiLog[];
-}
-
-export default function ApiLogSection({ logs }: ApiLogSectionProps) {
-  return <ApiLogPanel logs={logs} />;
-}
+export { default } from "../panels/ApiLogPanel";
