+++ mg_ui/mg_web_ui/frontend/src/components/status/ContainerStatusCard.tsx
@@ -1,0 +1,18 @@
+import SectionCard from "../layout/SectionCard";
+import StatusBadge from "../ui/StatusBadge";
+
+interface ContainerStatusCardProps {
+  title: string;
+  status: string;
+}
+
+export default function ContainerStatusCard({
+  title,
+  status,
+}: ContainerStatusCardProps) {
+  return (
+    <SectionCard title={title}>
+      <StatusBadge status={status} />
+    </SectionCard>
+  );
+}
