+++ mg_ui/mg_web_ui/frontend/src/components/ui/SectionLabel.tsx
@@ -1,0 +1,7 @@
+interface SectionLabelProps {
+  text: string;
+}
+
+export default function SectionLabel({ text }: SectionLabelProps) {
+  return <p className="text-xs text-gray-400">{text}</p>;
+}
