+++ mg_ui/mg_web_ui/frontend/src/ros/namespaces.ts
@@ -1,0 +1,9 @@
+export const NODE_NS = {
+  WAYPOINT_SEQUENCER: 'waypoint_sequencer_node',
+  DIAGNOSTICS: '',
+  LOCALIZATION: '',
+} as const
+
+export function nodeNs(ns: string, path: string): string {
+  return ns ? `/${ns}${path}` : path
+}
