+++ mg_ui/mg_web_ui/frontend/src/ros/topics.ts
@@ -1,0 +1,33 @@
+import { NODE_NS, nodeNs } from './namespaces'
+
+export const TOPICS = {
+  WAYPOINT_STATUS: nodeNs(NODE_NS.WAYPOINT_SEQUENCER, '/status'),
+  WAYPOINT_PAUSE_REQUEST: nodeNs(NODE_NS.WAYPOINT_SEQUENCER, '/pause_request'),
+  WAYPOINT_SET_NEXT_INDEX: nodeNs(NODE_NS.WAYPOINT_SEQUENCER, '/set_next_waypoint_index'),
+  DIAGNOSTICS: nodeNs(NODE_NS.DIAGNOSTICS, '/diagnostics'),
+  INITIALPOSE: nodeNs(NODE_NS.LOCALIZATION, '/initialpose'),
+  NAV_ACTION_STATUS: '/navigate_to_pose/_action/status',
+  AMCL_POSE: '/amcl_pose',
+  MAP: '/map',
+  TF: '/tf',
+  TF_STATIC: '/tf_static',
+  SCAN_TOP: '/scan_top_lidar',
+  SCAN_FRONT: '/scan_front_lidar',
+  NAV_PLAN: '/plan',
+  ACTUAL_PATH: '/actual_path',
+  GLOBAL_COSTMAP: '/global_costmap/costmap',
+  LOCAL_COSTMAP: '/local_costmap/costmap',
+  WAYPOINT_MARKERS: '/waypoint_sequencer_node/waypoints_markers',
+  COLLISION_FRONT: '/collision_detector/polygon_front',
+  COLLISION_REAR: '/collision_detector/polygon_rear',
+  DEPTH_POINTS: '/rs_d435i/depth/color/points',
+  DEPTH_IMAGE: '/rs_d435i/depth/image_raw',
+  CAMERA_IMAGE: '/rs_d435i/color/image_raw',
+  PARTICLE_CLOUD: '/particle_cloud',
+  CMD_VEL: '/cmd_vel',
+  ODOM: '/odom',
+  EMERGENCY_STOP: '/motor_driver_node/emergency_stop',
+  COLLISION_STATE: '/collision_detector_state',
+  CPU_USAGE: '/cpu_usage',
+  MEMORY_USAGE: '/memory_usage',
+} as const
