+++ mg_ui/mg_web_ui/frontend/src/types/api.ts
@@ -1,0 +1,20 @@
+export type ContainerStatus = 'running' | 'exited' | 'dead' | 'unknown'
+
+export type CallApi = (
+  path: string,
+  body?: unknown,
+) => Promise<{ success: boolean; message: string }>
+
+export interface ApiLog {
+  id: number
+  timestamp: string
+  path: string
+  success: boolean
+  message: string
+}
+
+export interface SystemManagerHandle {
+  containers: Record<string, string>
+  logs: ApiLog[]
+  callApi: CallApi
+}
